pyproject.toml
[tool.poetry]
name = "gujpol-slm"
version = "0.1.0"
description = "Gujarat Police AI-Powered Investigation Support System"
authors = ["Gujarat Police AI Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
python-multipart = "^0.0.12"

# Database
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"

# Vector Database
chromadb = "^0.5.0"

# Redis
redis = "^5.1.0"
celery = "^5.4.0"

# ML / NLP
sentence-transformers = "^3.1.0"
transformers = "^4.45.0"
torch = "^2.4.0"
peft = "^0.13.0"
bitsandbytes = "^0.44.0"
trl = "^0.11.0"
datasets = "^3.0.0"
accelerate = "^0.34.0"

# OCR
pytesseract = "^0.3.13"
paddleocr = "^2.8.0"
paddlepaddle = "^2.6.0"
Pillow = "^10.4.0"
pdf2image = "^1.17.0"

# Document Processing
python-docx = "^1.1.0"
PyPDF2 = "^3.0.0"
pdfplumber = "^0.11.0"
python-magic = "^0.4.27"
chardet = "^5.2.0"

# Data Processing
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.5.0"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.25.0"
langdetect = "^1.0.9"
indic-transliteration = "^2.3.0"

# Scraping (for verified legal data sources)
requests = "^2.32.0"
beautifulsoup4 = "^4.12.0"
scrapy = "^2.11.0"
selenium = "^4.25.0"
lxml = "^5.3.0"
aiohttp = "^3.10.0"

# Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "^43.0.0"

# Monitoring & Logging
prometheus-fastapi-instrumentator = "^7.0.0"
structlog = "^24.4.0"

# Utilities
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
click = "^8.1.7"
tqdm = "^4.66.0"
tenacity = "^9.0.0"
schedule = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
factory-boy = "^3.3.0"
faker = "^30.0.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
pre-commit = "^3.8.0"

[tool.poetry.scripts]
gujpol = "src.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
